Update fr - alternate unionfaqs.md
We realized that there are some problems with "what's a game worker" unionfaqs article:
The latest english version on the website was re-written in 2022, and it's longer. The lastest english zine layout however is not using this new longer version.
That updated english version was re-translated in 2022, not based on the original french translation from 2019. That new french version is much longer than the latest english version, which is even more of a problem for fitting into the layout. 
So we're writing a new french version, pulling from both the original 2019 version and the 2022 rewrite, aiming for an equivalent length to the latest english zine (not website) version.